Reorganize parts directory
