+++ nemo/collections/asr/data/audio_to_text.py
@@ -26,7 +26,7 @@ from torch.nn import functional as F
 
 from nemo.collections.asr.data import vocabs
 from nemo.collections.common.parts.preprocessing import collections, parsers
-from nemo.collections.asr.parts.features import WaveformFeaturizer
+from nemo.collections.asr.parts.preprocessing.features import WaveformFeaturizer
 from nemo.core.classes import Dataset, IterableDataset
 from nemo.core.neural_types import *
 from nemo.core.neural_types.elements import ProbsType
