+++ nemo/collections/asr/modules/conv_asr.py
@@ -20,7 +20,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 from omegaconf import MISSING, ListConfig, OmegaConf
 
-from nemo.collections.asr.parts.jasper import (
+from nemo.collections.asr.parts.submodules.jasper import (
     JasperBlock,
     MaskedConv1d,
     StatsPoolLayer,
