+++ nemo/collections/asr/modules/rnnt.py
@@ -31,7 +31,7 @@ from typing import Any, Dict, List, Optional, Tuple, Union
 import torch
 
 from nemo.collections.asr.modules import rnnt_abstract
-from nemo.collections.asr.parts import rnnt_utils
+from nemo.collections.asr.parts.utils import rnnt_utils
 from nemo.collections.common.parts import rnn
 from nemo.core.classes import typecheck
 from nemo.core.neural_types import (
