+++ nemo/collections/asr/parts/submodules/conformer_modules.py
@@ -16,8 +16,8 @@ import torch
 from torch import nn as nn
 from torch.nn import LayerNorm
 
-from nemo.collections.asr.parts.activations import Swish
-from nemo.collections.asr.parts.multi_head_attention import MultiHeadAttention, RelPositionMultiHeadAttention
+from nemo.collections.asr.parts.utils.activations import Swish
+from nemo.collections.asr.parts.submodules.multi_head_attention import MultiHeadAttention, RelPositionMultiHeadAttention
 
 __all__ = ['ConformerConvolution', 'ConformerFeedForward', 'ConformerLayer']
