+++ nemo/collections/asr/parts/submodules/multi_head_attention.py
