+++ scripts/voice_activity_detection/vad_tune_threshold.py
@@ -16,7 +16,7 @@ import argparse
 
 import numpy as np
 
-from nemo.collections.asr.parts.vad_utils import vad_tune_threshold_on_dev
+from nemo.collections.asr.parts.utils.vad_utils import vad_tune_threshold_on_dev
 from nemo.utils import logging
 
 if __name__ == "__main__":
