+++ tests/collections/asr/test_asr_filterbankfeatures_seq_len.py
@@ -17,7 +17,7 @@ import numpy as np
 import pytest
 import torch
 
-from nemo.collections.asr.parts.features import FilterbankFeatures
+from nemo.collections.asr.parts.preprocessing.features import FilterbankFeatures
 
 
 class TestFilterbankFeatures:
