+++ tests/collections/asr/test_asr_rnnt_encdec_model.py
@@ -17,10 +17,9 @@ import pytest
 import torch
 from omegaconf import DictConfig, ListConfig
 
-from nemo.collections.asr.metrics import rnnt_wer
 from nemo.collections.asr.models import EncDecRNNTModel
-from nemo.collections.asr.parts import rnnt_beam_decoding as beam_decode
-from nemo.collections.asr.parts import rnnt_greedy_decoding as greedy_decode
+from nemo.collections.asr.parts.submodules import rnnt_greedy_decoding as greedy_decode, \
+    rnnt_beam_decoding as beam_decode
 from nemo.collections.asr.parts.numba import __NUMBA_MINIMUM_VERSION__, numba_utils
 from nemo.utils.config_utils import assert_dataclass_signature_match
